BC-662 | Extracted invalid date arguments to reusable provider for validator tests
