+++ src/test/java/uk/gov/justice/laa/amend/claim/forms/amendments/validators/unique/CivilCaseConcludedDateValidatorTest.java
@@ -4,11 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Map;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.Errors;
@@ -33,33 +31,9 @@ class CivilCaseConcludedDateValidatorTest {
     assertTrue(validator.appliesTo(CivilClaimDetailsViewField.CASE_CONCLUDED_CLAIMED_DATE));
   }
 
-  static Stream<Arguments> invalidDateProvider() {
-    return Stream.of(
-        // Empty strings
-        Arguments.of("", "", ""),
-        Arguments.of("", "12", "2020"),
-        Arguments.of("15", "", "2020"),
-        Arguments.of("15", "12", ""),
-        // Invalid day (40th date)
-        Arguments.of("40", "12", "2020"),
-        Arguments.of("32", "01", "2020"),
-        Arguments.of("00", "12", "2020"),
-        // Invalid month
-        Arguments.of("15", "13", "2020"),
-        Arguments.of("15", "00", "2020"),
-        // Characters instead of numbers
-        Arguments.of("abc", "12", "2020"),
-        Arguments.of("15", "xyz", "2020"),
-        Arguments.of("15", "12", "abcd"),
-        Arguments.of("!@#", "$%^", "&*()"),
-        // Mixed valid and invalid characters
-        Arguments.of("1a", "12", "2020"),
-        Arguments.of("15", "1b", "2020"),
-        Arguments.of("15", "12", "20c0"));
-  }
-
   @ParameterizedTest
-  @MethodSource("invalidDateProvider")
+  @MethodSource(
+      "uk.gov.justice.laa.amend.claim.forms.amendments.validators.InvalidDateArgumentsProvider#invalidDateProvider")
   void shouldNotAddErrorsForInvalidDates(String day, String month, String year) {
     Map<String, String> input =
         Map.of(
